+++ src/generals/RandomShooterGeneral.sol
@@ -18,7 +18,7 @@ contract RandomShooterGeneral is IGeneral {
     }
 
     function fire(
-        uint192, /* myBoard */
+        uint256, /* myBoard */
         uint192 myAttacks,
         uint192, /* opponentsAttacks */
         uint256 myLastMove,
